renamed algorithm file names and fixec icon issue for mac
